[PIR-1513] Database Query is not cancelled when hitting 'Cancel' during the run/build of an Interactive Report
This code is for HotFix the cancel query. The listener has info of status but did not have a way to ask if the report was cancelled. added this to simplify the question by lower level items.

NOTE: since this ListenerStatus is available via LocalThread, then any class within the correct Thread will be able to see the correct value for their execution.